Rename FeatureLengthAxis to BasePairsRuler
