+++ src/ensembl/src/content/app/entity-viewer/components/base-pairs-ruler/BasePairsRuler.scss
