+++ src/ensembl/src/content/app/entity-viewer/components/base-pairs-ruler/BasePairsRuler.test.tsx
@@ -1,24 +1,24 @@
 import React from 'react';
 import { mount, render } from 'enzyme';
 
-import FeatureLengthAxis from './FeatureLengthAxis';
+import BasePairsRuler from './BasePairsRuler';
 
 const defaultProps = {
   length: 80792,
   width: 600
 };
 
-describe('<FeatureLengthAxis />', () => {
+describe('<BasePairsRuler />', () => {
   describe('rendering', () => {
     it('renders inside an <svg> element if standalone', () => {
       const wrapper = render(
-        <FeatureLengthAxis {...defaultProps} standalone={true} />
+        <BasePairsRuler {...defaultProps} standalone={true} />
       );
       expect(wrapper.is('svg')).toBe(true);
     });
 
     it('renders inside a <g> element (svg group) if not standalone', () => {
-      const wrapper = render(<FeatureLengthAxis {...defaultProps} />);
+      const wrapper = render(<BasePairsRuler {...defaultProps} />);
       expect(wrapper.is('g')).toBe(true);
     });
   });
@@ -31,7 +31,7 @@ describe('<FeatureLengthAxis />', () => {
 
     it('passes calculated ticks to the callback', () => {
       const callback = jest.fn();
-      mount(<FeatureLengthAxis {...props} onTicksCalculated={callback} />);
+      mount(<BasePairsRuler {...props} onTicksCalculated={callback} />);
       expect(callback).toHaveBeenCalledTimes(1);
 
       const payload = callback.mock.calls[0][0];
