+++ src/ensembl/src/content/app/entity-viewer/components/base-pairs-ruler/basePairsRulerHelper.test.ts
@@ -1,6 +1,6 @@
 import { scaleLinear } from 'd3';
 
-import { getTicks } from './featureLengthAxisHelper';
+import { getTicks } from './basePairsRulerHelper';
 
 type Example = {
   length: number;
