+++ src/ensembl/stories/entity-viewer/base-pairs-ruler/BasePairsRuler.stories.scss
